+++ src/main/java/raisetech/Student_Management/StudentManagementApplication.java
@@ -1,32 +1,13 @@
 package raisetech.Student_Management;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
 
-	@Autowired
-	private StudentRepository repository;
 
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
-	}
-
-	@GetMapping("/studentList")
-	public List<Student> getStudentList() {
-		return repository.search();
-	}
-
-//	課題用
-	@GetMapping("studentsCourseList")
-	public List<StudentsCourses> getStudentsCourseList() {
-		return repository.searchStudentCourses();
 	}
 }
