+++ src/main/java/raisetech/Student_Management/controller/StudentController.java
@@ -1,0 +1,32 @@
+package raisetech.Student_Management.controller;
+
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.Student_Management.data.Student;
+import raisetech.Student_Management.data.StudentsCourses;
+import raisetech.Student_Management.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  //	課題用
+  @GetMapping("/studentsCourseList")
+  public List<StudentsCourses> getStudentsCourseList() {
+    return service.searchStudentsCourseList();
+  }
+}
