+++ src/main/java/raisetech/Student_Management/repository/StudentRepository.java
@@ -1,9 +1,12 @@
-package raisetech.Student_Management;
+package raisetech.Student_Management.repository;
 
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.Student_Management.data.Student;
+import raisetech.Student_Management.data.StudentsCourses;
+
 
 @Mapper
 public interface StudentRepository {
@@ -11,7 +14,6 @@ public interface StudentRepository {
   @Select("SELECT * FROM students")
   List<Student> search();
 
-//  課題用
   @Select("SELECT * FROM students_courses")
   List<StudentsCourses> searchStudentCourses();
 }
